이미지가 없는 굿즈 삭제시 삭제 실패 오류 해결

+++ src/main/java/com/walking/project_walking/service/GoodsService.java
@@ -54,7 +54,8 @@ public class GoodsService {
   @Transactional
   public Boolean deleteGoods(Long id) {
     GoodsResponseDto goods = getGoodsById(id);
-    imageService.deleteImageFromS3(goods.getGoodsImage());
+    if (goods.getGoodsImage() != null)
+      imageService.deleteImageFromS3(goods.getGoodsImage());
     if (goods != null) {
       goodsRepository.deleteById(id);
       return Boolean.TRUE;
